pyproject.toml
[tool.poetry]
name = "news-app"
version = "0.1.0"
description = ""
authors = ["tikendraw <tikendraksahu1029@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
icecream = "^2.1.3"
fastapi = "^0.110.1"
sqlalchemy-cockroachdb = "^2.0.2"
pydantic = "^2.6.4"
flask = "^3.0.2"
requests-cache = "^1.2.0"
pre-commit = "^3.7.0"
python-dotenv = "^1.0.1"
newsdataapi = "^0.1.13"
bs4 = "^0.0.2"
py-pipelines = "^1.0.3"
langchain = "^0.1.14"
google-generativeai = "^0.4.1"
transformers = {extras = ["sentencepiece"], version = "^4.39.3"}
datasets = "^2.18.0"
sacrebleu = "^2.4.1"
rouge-score = "^0.1.2"
py7zr = "^0.21.0"
pandas = "^2.2.1"
nltk = "^3.8.1"
tqdm = "^4.66.2"
pyyaml = "^6.0.1"
matplotlib = "^3.8.4"
notebook = "^7.1.2"
ipykernel = "^6.29.4"
boto3 = "^1.34.78"
mypy-boto3-s3 = "^1.34.65"
langchain-google-genai = "^1.0.1"

[tool.poetry.group.dev.dependencies]
fake-useragent = "^1.5.1"
aiohttp = "^3.9.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
